Camview partially working
